+++ views/control_structures_dockwidget.py
@@ -1,0 +1,45 @@
+# -*- coding: utf-8 -*-
+import os
+import logging
+
+from PyQt4 import uic
+from PyQt4 import QtCore
+from PyQt4 import QtGui
+from PyQt4.QtCore import pyqtSignal
+from PyQt4.QtGui import QApplication
+
+log = logging.getLogger(__name__)
+
+try:
+    _fromUtf8 = QtCore.QString.fromUtf8
+except AttributeError:
+    def _fromUtf8(s):
+        return s
+
+try:
+    _encoding = QApplication.UnicodeUTF8
+
+    def _translate(context, text, disambig):
+        return QApplication.translate(context, text, disambig, _encoding)
+except AttributeError:
+    def _translate(context, text, disambig):
+        return QApplication.translate(context, text, disambig)
+
+FORM_CLASS, _ = uic.loadUiType(os.path.join(
+    os.path.dirname(__file__), '..', 'ui',
+    'control_structures_dockwidget.ui'))
+
+
+class ControlStructuresDockWidget(QtGui.QDockWidget, FORM_CLASS):
+
+    closingWidget = pyqtSignal()
+
+    def __init__(self, parent=None):
+        """Constructor."""
+        super(ControlStructuresDockWidget, self).__init__(parent)
+        # Set up the user interface from Designer.
+        self.setupUi(self)
+
+    def closeEvent(self, event):
+        self.closingWidget.emit()
+        event.accept()